[move-ide] Added a skeleton of the Prettier Move plugin (#15691)
## Description 

This PR adds a skeleton of Move plugin for
[Prettier](https://github.com/prettier/prettier). It only supports a few
constructs at the moment but it has been verified that it works both on
the command line and in VSCode, and it's platform independent due to it
using a WASM-ified version of tree-sitter's Move parser.

## Test Plan 

Manually tested that it works on the command line and with VSCode. Added
a testing framework to test individual features and added tests for
existing features.