[move-ide] Refactoring to reduce memory footprint and prepare for new features (#16116)
## Description 

This is mainly a refactoring with a small change to how constants are
displayed (to include their defining module).

The refactoring started while attempting to implement a new feature but
it was deemed a better idea to create a separate PR for it. In addition
to enabling a new feature, it also reduces memory consumption by storing
additional information for definitions on the side instead of inlining
them in the `UseDef` structure - we moved `IdentOnHover` (now called
`DefInfo` as it will be used for other things as well) to a separate
map.


## Test Plan 

All existing tests (after new constant display adjustment) pass.