[consensus] Add universal committer tests (#15923)
Most test scenarios are identical to the test in base_committer_tests.rs
but just going through the `UniversalCommitter` logic to handle checking
between direct & indirect rules. Feel free to review all tests but the
new tests that were not relevant in base_committer_tests.rs are
- `direct_commit_late_call`
- `no_genesis_commit`
- `no_leader`

Will follow up in another PR with the tests for multi leader and
pipelining.