toolchain versioning: fix windows binary rename (#16182)
## Description 

On windows we have a step to rename the compiler binary like
`sui-windows-x86_64` to `sui`. Except that, the windows release tarball
names the source binary `sui-windows-x86_64.exe`, so we need to append
the `.exe` for the rename to succeed.

Part of end-to-end testing on windows and ran into this.

## Test Plan 

Tested the fix with local source build--we don't have windows CI env for
testing (yet?) so this will have to do for now!

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes