[GraphQL] fix issue with failed dry run and add more e2e test (#16248)
## Description 

Currently we assume that the dev inspect result coming from fullnode is
always `Some` but it can actually be `None` if the transaction failed to
execute. This PR fixes that and adds a few more e2e test cases for this
scenario as well as a scenario where a `TransactionKind` is provided.

## Test Plan 

Added tests

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes