[NameService] Adds support to more input formats (#15466)
## Description 

**Updated the description (16-01-2024) due to a change in our support
(we won't omit TLDs and `@` is a special format after all, not another
wildcard)**

With this PR, we add support for wildcard `.` or `*` as the separator
for a valid domain, and also support the new SuiNS display format, based
on `@`.

Some examples:
`@test` -> `test.sui`
`sub@test` -> `sub.test.sui`
`inner.sub@test` -> `inner.sub.test.sui`
`more.inner.sub@test` -> `more.inner.sub.test.sui`

All of the above would work with `*` instead of `.` too.

## Test Plan 

Added a new test file (We didn't have one for the name service). Not
sure if this is the right way to test stuff on our crates, so please
advise me here if it's not right.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes