[Consensus 2.0] re-broadcast last proposed block after recover (#16751)
## Description 

To ensure liveness we should attempt to re-broadcast the last proposed
block after recovery if no new one created. This is a fix for the case
but a more proper solution should be provided as discussed in the past
to ensure re-broadcast of last proposed block under no new round
advancement

## Test Plan 

CI

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes