fix: Graceful shutdown of pool workers in data ingestion framework (#16814)
## Description 

In the data ingestion workers, when send channels return `Err` it is a
sign the processing pipeline is being closed, but right now this can
lead to panics due to the use of `expects`. This PR replaces these with
break or return statements, and ensures that all workers are finished
before returning.

## Test Plan 

Unit tests are run.