feat: add upper bound for max epoch in zklogin sig (#16233)
## Description 

currently a zklogin sig is valid as long as max_epoch is larger than
current epoch. this adds a protocol config that the max_epoch cannot be
too large, the config sets the upper bound to 30 + current epoch.

based on new testing framework in simtest only. 

## Test Plan 

simtest
---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [x] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Protocol version upgrades to 42. This adds a feature flag to set the
upper bound of the max epoch for a zklogin signature.