Add --chain option to sui-tool replay commands (#17187)
## Description 

When rpc_url is not specified, previously the sui-tool would iterate
through the network config and try each config one by one. This is
rather unnecessary and inefficient.
This PR adds an explicit --chain option so that when rpc_url is not
specified, it will use the config named `chain` in the config file. This
simplifies some of the implementation and makes it faster to replay a
transaction when the desired config is not the first one in the config
file.

## Test plan 

CI
Run locally

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: