Re-execute locally built checkpoints at startup (#16973)
This will be necessary for the WritebackCache, and unnecessary but
harmless otherwise.

Until built checkpoints are re-executed, the validator is an in
inconsistent state, since
every locally checkpointed transaction should have effects locally
available. Without this change, we hit this case
https://github.com/MystenLabs/sui/blob/59168f73ffd71023de6427186e5dac206cf4698e/crates/sui-core/src/storage.rs#L172
when we shouldn't.

This is a temporary measure. Once we cache and quarantine consensus
output, we will simply re-process any consensus commits which have not
yet resulted in a certified checkpoint, so we won't have this edge case
anymore.