Ensure that reconfiguration cannot happen while we are in enqueue_impl (or other methods) (#17211)
In enqueue_impl, we acquire inner.write() twice - in between these two
critical sections we need to ensure that the cache is in unbounded mode
so that we do not miss object writes that happen in between.

However, since we do not hold the lock the entire time, reconfiguration
could previously happen. This would reset the inner state, and then we
could crash at
https://github.com/MystenLabs/sui/blob/d181e3f6ec8c3e185c239543f2613103db806eee/crates/sui-core/src/transaction_manager.rs#L201
when attempting to disable unbounded mode.

This PR ensures that reconfig cannot happen at any point during
`enqueue_impl` or any other method that holds the lock. (Note that other
methods were not affected since `enqueue_impl` is the only one that has
two separate critsecs)